# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job Matching & Outreach API",
    "description": "Scrape job listings, match them to a resume with embeddings, and draft tailored cold emails with an LLM.",
    "version": "0.2.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/extract": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Extract",
        "description": "Scrape a careers page and extract structured job postings.",
        "operationId": "extract_api_v1_jobs_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resume/parse": {
      "post": {
        "tags": [
          "resume"
        ],
        "summary": "Parse Resume",
        "description": "Extract text from a PDF resume.",
        "operationId": "parse_resume_api_v1_resume_parse_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_resume_api_v1_resume_parse_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/match": {
      "post": {
        "tags": [
          "matching"
        ],
        "summary": "Match",
        "description": "Rank the jobs on a careers page against a resume.",
        "operationId": "match_api_v1_match_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_match_api_v1_match_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/emails": {
      "post": {
        "tags": [
          "email"
        ],
        "summary": "Email",
        "description": "Write a cold email for one job, grounded in the resume text.",
        "operationId": "email_api_v1_emails_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pipeline": {
      "post": {
        "tags": [
          "pipeline"
        ],
        "summary": "Pipeline",
        "description": "End to end: scrape \u2192 extract \u2192 match \u2192 draft an email for the best match.",
        "operationId": "pipeline_api_v1_pipeline_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_pipeline_api_v1_pipeline_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_match_api_v1_match_post": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Careers / job listings page URL"
          },
          "resume": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Resume",
            "description": "Resume as a PDF"
          },
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "url",
          "resume"
        ],
        "title": "Body_match_api_v1_match_post"
      },
      "Body_parse_resume_api_v1_resume_parse_post": {
        "properties": {
          "resume": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Resume",
            "description": "Resume as a PDF"
          }
        },
        "type": "object",
        "required": [
          "resume"
        ],
        "title": "Body_parse_resume_api_v1_resume_parse_post"
      },
      "Body_pipeline_api_v1_pipeline_post": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Careers / job listings page URL"
          },
          "resume": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Resume",
            "description": "Resume as a PDF"
          },
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          },
          "candidate_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Name"
          }
        },
        "type": "object",
        "required": [
          "url",
          "resume"
        ],
        "title": "Body_pipeline_api_v1_pipeline_post"
      },
      "EmailRequest": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/Job"
          },
          "resume_text": {
            "type": "string",
            "minLength": 50,
            "title": "Resume Text"
          },
          "candidate_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Name"
          }
        },
        "type": "object",
        "required": [
          "job",
          "resume_text"
        ],
        "title": "EmailRequest"
      },
      "EmailResponse": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "EmailResponse"
      },
      "ExtractRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ExtractRequest"
      },
      "ExtractResponse": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/Job"
            },
            "type": "array",
            "title": "Jobs"
          }
        },
        "type": "object",
        "required": [
          "url",
          "count",
          "jobs"
        ],
        "title": "ExtractResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "llm_configured": {
            "type": "boolean",
            "title": "Llm Configured"
          }
        },
        "type": "object",
        "required": [
          "status",
          "version",
          "llm_configured"
        ],
        "title": "HealthResponse"
      },
      "Job": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role",
            "default": ""
          },
          "experience": {
            "type": "string",
            "title": "Experience",
            "default": ""
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "title": "Job"
      },
      "Match": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "score": {
            "type": "number",
            "title": "Score",
            "description": "Cosine similarity between resume and job (higher is better)"
          },
          "job": {
            "$ref": "#/components/schemas/Job"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "score",
          "job"
        ],
        "title": "Match"
      },
      "MatchResponse": {
        "properties": {
          "jobs_found": {
            "type": "integer",
            "title": "Jobs Found"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/Match"
            },
            "type": "array",
            "title": "Matches"
          }
        },
        "type": "object",
        "required": [
          "jobs_found",
          "matches"
        ],
        "title": "MatchResponse"
      },
      "PipelineResponse": {
        "properties": {
          "jobs_found": {
            "type": "integer",
            "title": "Jobs Found"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/Match"
            },
            "type": "array",
            "title": "Matches"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "jobs_found",
          "matches"
        ],
        "title": "PipelineResponse"
      },
      "ResumeResponse": {
        "properties": {
          "chars": {
            "type": "integer",
            "title": "Chars"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "chars",
          "text"
        ],
        "title": "ResumeResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}